clean: Rename dataflow script
